Qvectors per spin now supported
-Q qfile
can be given in the command line of findAngles
or montecarlo to indicate a file from where to
read the moments for the spins.
The format of the file is

\#version
N 3
q00 q01 q02
q10 q11 q12
...

where N is the number of spins.
Spaces and end-of-line are equivalent
in this file.

The use of -q q0,q1,q2 is still supported,
but obviously cannot be used at the same
time as -Q file is used.